backend/requirements.txt
fastapi
uvicorn[standard]
pandas
anthropic
python-multipart
python-dotenv
pydantic-settings
google-genai
openai
pytest
httpx
